Refactoring, SOLID, Flamethrower doing percentage of health damage correctly implemented

+++ src/main/java/com/group9/model/entities/characters/attackers/AttackerType.java
@@ -6,7 +6,7 @@ public class AttackerType extends CharacterType {
     private int resourceReward;
     private int speed;
 
-    public AttackerType(String name, int maxHealth, int attackDamage, int range, float attackDelay, int speed, int resourceReward) {
+    public AttackerType(String name, float maxHealth, float attackDamage, int range, float attackDelay, int speed, int resourceReward) {
         super(name, maxHealth, attackDamage, range, attackDelay);
         this.resourceReward = resourceReward;
         this.speed = speed;
