+++ src/main/java/com/group9/model/entities/characters/defenders/DefenderType.java
@@ -8,7 +8,7 @@ public class DefenderType extends CharacterType {
     private final int cost;
     private final AttackStrategy attackStrategy;
 
-    public DefenderType(String name, int maxHealth, int attackDamage, int range, float attackDelay, int cost, AttackStrategy attackStrategy) {
+    public DefenderType(String name, float maxHealth, float attackDamage, int range, float attackDelay, int cost, AttackStrategy attackStrategy) {
         super(name, maxHealth, attackDamage, range, attackDelay);
         this.cost = cost;
         this.attackStrategy = attackStrategy;
